Support Go 1.16+ compat
