+++ compress/flate/inflate_test.go
@@ -115,21 +115,3 @@ func opticks(t testing.TB) (data []byte) {
 	}
 	return data
 }
-
-func FuzzInflate(f *testing.F) {
-	data := opticks(f)
-	f.Add(data)
-	f.Fuzz(func(t *testing.T, source []byte) {
-		input := compress(source)
-		r := NewReader(bytes.NewReader(input))
-		var err error
-		data, err := io.ReadAll(r)
-		n := len(data)
-		if err != nil && err != io.EOF {
-			t.Fatal(err, n, bytes.Equal(data[:n], source[:n]))
-		}
-		if !bytes.Equal(data[:n], source) {
-			t.Fatal()
-		}
-	})
-}
