+++ compress/flate/internal/deflate/huffmanonly_amd64.s
@@ -1,6 +1,7 @@
 // Code generated by command: go run main.go. DO NOT EDIT.
 
 //go:build amd64 && !noasmtest
+// +build amd64,!noasmtest
 
 #include "textflag.h"
