+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/intel/fastgo
 
-go 1.20
+go 1.16
