+++ internal/cpu/detect_amd64.s
@@ -1,5 +1,6 @@
 // Code generated by command: go run main.go. DO NOT EDIT.
 //go:build amd64 && !noasmtest
+// +build amd64,!noasmtest
 
 #include "textflag.h"
