+++ internal/cpu/level.go
@@ -1,8 +1,6 @@
 // Copyright (c) 2024, Intel Corporation.
 // SPDX-License-Identifier: BSD-3-Clause
 
-//go:build amd64 && !noasmtest
-
 package cpu
 
 // check arch level while loading package
